requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
ollama==0.1.7
html5lib==1.1
